+++ clang/lib/Sema/SemaTemplateInstantiate.cpp
@@ -1420,9 +1420,9 @@ namespace {
     TransformIntelBlockLoopAttr(const IntelBlockLoopAttr *BL);
     const IntelPrefetchAttr *
     TransformIntelPrefetchAttr(const IntelPrefetchAttr *IPA);
-    const CodeAlignAttr *TransformCodeAlignAttr(const CodeAlignAttr *CA);
 #endif // INTEL_CUSTOMIZATION
 
+    const CodeAlignAttr *TransformCodeAlignAttr(const CodeAlignAttr *CA);
     ExprResult TransformPredefinedExpr(PredefinedExpr *E);
     ExprResult TransformDeclRefExpr(DeclRefExpr *E);
     ExprResult TransformCXXDefaultArgExpr(CXXDefaultArgExpr *E);
@@ -2131,13 +2131,11 @@ TemplateInstantiator::TransformSYCLIntelMaxReinvocationDelayAttr(
                                                               TransformedExpr);
 }
 
-#if INTEL_CUSTOMIZATION
 const CodeAlignAttr *
 TemplateInstantiator::TransformCodeAlignAttr(const CodeAlignAttr *CA) {
   Expr *TransformedExpr = getDerived().TransformExpr(CA->getAlignment()).get();
   return getSema().BuildCodeAlignAttr(*CA, TransformedExpr);
 }
-#endif // INTEL_CUSTOMIZATION
 
 ExprResult TemplateInstantiator::transformNonTypeTemplateParmRef(
     Decl *AssociatedDecl, const NonTypeTemplateParmDecl *parm,