+++ clang/lib/Sema/TreeTransform.h
@@ -1403,7 +1403,8 @@ public:
   StmtResult RebuildAttributedStmt(SourceLocation AttrLoc,
                                    ArrayRef<const Attr *> Attrs,
                                    Stmt *SubStmt) {
-    if (SemaRef.CheckRebuiltAttributedStmtAttributes(Attrs))
+    if (SemaRef.CheckRebuiltAttributedStmtAttributes(Attrs) ||
+        SemaRef.CheckRebuiltCodeAlignStmtAttributes(Attrs))
       return StmtError();
     return SemaRef.BuildAttributedStmt(AttrLoc, Attrs, SubStmt);
   }